requirements.txt
alembic

astroid

certifi

chardet

click

colorama

Flask

Flask-Login

flask-marshmallow

Flask-Migrate

Flask-Script

Flask-SQLAlchemy

idna

isort

itsdangerous

Jinja2

lazy-object-proxy

Mako

MarkupSafe

marshmallow

marshmallow-sqlalchemy

mccabe

psycopg2-binary

pylint

python-dateutil

python-dotenv

python-editor

requests

six

SQLAlchemy

toml

urllib3

Werkzeug

wrapt

